requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudscraper==1.2.71
colorama==0.4.6
h11==0.16.0
idna==3.10
MouseInfo==0.1.3
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0